+++ src/components/Home/Content/Construction.js
@@ -1,0 +1,14 @@
+import React from 'react';
+import HomeCountdown from '../../Countdown';
+
+const Construction = () => {
+    return(
+        <div className="col-sm-12">
+            <h2>Under Construction</h2>
+            <p>We are working very hard on the new version of our site. It will bring a lot of new features. Stay tuned!</p>
+            <HomeCountdown/>
+        </div>
+    )
+}
+
+export default Construction;
