+++ src/components/Home/Content/index.js
@@ -1,0 +1,14 @@
+import React from 'react';
+import ContentConstruction from './Construction';
+
+const Content = () => {
+    return (
+        <div className="container">
+            <div className="row">
+                <ContentConstruction/>
+            </div>
+        </div>
+    )
+}
+
+export default Content;
