+++ src/screens/Footer/ContactForm.js
@@ -1,0 +1,14 @@
+import React from 'react';
+import Contact from '../../components/Contact';
+
+const ContactForm = () => {
+    return(
+        <div className="col-sm-12 subscribe">
+                    <h3>Your suggestions means a lot</h3>
+                    <p>Send us your query and we'll get you back ASAP</p>                    
+                    <Contact/>
+                </div>
+    )
+}
+
+export default ContactForm;
