+++ src/screens/Header/Info.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const Info = () => {
+    return (
+        <div class="col-sm-8 call-us">
+                    <p>Tel: <span>0039 123 45 789</span> | Skype: <span>info@imaginetech.net</span></p>
+                </div>
+    )
+}
+
+export default Info;
