+++ src/screens/Header/Logo.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const Logo = () => {
+    return (
+        <div class="col-sm-4 logo">
+                    <h1><a href="#">ImagineTech</a> <span>.</span></h1>
+        </div>
+    )
+}
+
+export default Logo;
