+++ src/screens/Header/index.js
@@ -1,16 +1,14 @@
 import React, { Component } from 'react';
+import HeaderLogo from './Logo';
+import HeaderInfo from './Info';
 
 class HeaderContainer extends Component {
     render() {
         return (
             <div class="container">
             <div class="row header">
-                <div class="col-sm-4 logo">
-                    <h1><a href="index.html">ImagineTech</a> <span>.</span></h1>
-                </div>
-                <div class="col-sm-8 call-us">
-                    <p>Tel: <span>0039 123 45 789</span> | Skype: <span>info@imaginetech.net</span></p>
-                </div>
+                <HeaderLogo/>
+                <HeaderInfo/>
             </div>
             </div>
         );
